refactor(runtime): split runtime.rs into focused modules (tracker #011)
Decomposes the ~1.9k-line single-impl runtime.rs into 7 files along
natural seams, each adding its own `impl RunTime` block:

  runtime.rs (core, 353)  RunTime/ScopeFrame/IdHasher/VarMap/ScopeEvent,
                          new/register_semantic_func/alloc_str/resolve_str,
                          the From impls, and the shared free helpers.
  scope.rs   (385)        scope stack + variable storage (push/pop/declare,
                          get/set/container/array, write_var_at).
  ops.rs     (249)        apply_op, apply_unary.
  eval.rs    (151)        eval_semantic_expr, semantic_value_to_runtime.
  exec.rs    (532)        run_semantic_stmt, run_semantic_when.
  call.rs    (370)        call_semantic_func, call_semantic_method.
  print.rs   (61)         print_value(_inline), expand_interpolation.

Pure move — method bodies are byte-identical; the only non-move edits are
visibility (private fns called across the new boundaries became
pub(crate): apply_op, apply_unary, run_semantic_when, call_semantic_method,
print_value, print_value_inline, semantic_value_to_runtime, the ScopeFrame
helpers, the free helper fns, and the RunTime.scopes field). RunTime lives
in runtime.rs so `crate::runtime::runtime::{RunTime, ScopeEvent}` imports
(main.rs, diagnostics.rs) are untouched. No behavior change.

Closes tracker item #011.
