+++ src/runtime/mod.rs
@@ -1,2 +1,8 @@
 pub mod handle;
 pub mod runtime;
+pub mod scope;
+pub mod ops;
+pub mod eval;
+pub mod exec;
+pub mod call;
+pub mod print;
