+++ src/runtime/runtime.rs
@@ -1,6 +1,5 @@
 // incremental rebuild test 3
-use crate::frontend::{ast::*, diagnostics, types::*};
-use crate::frontend::builtins::{self, BuiltinKind};
+use crate::frontend::{ast::*, types::*};
 use crate::frontend::semantic_types::*;
 use crate::runtime::handle::HandleRegistry;
 use std::collections::{HashMap, HashSet};
@@ -63,7 +62,7 @@ pub struct ScopeFrame {
 }
 
 impl ScopeFrame {
-    fn new() -> Self {
+    pub(crate) fn new() -> Self {
         ScopeFrame {
             vars: VarMap::default(),
             by_name: HashMap::new(),
@@ -74,23 +73,23 @@ impl ScopeFrame {
     }
 
     /// Resolve a variable entry in this frame by name (cold path).
-    fn get_by_name(&self, name: &str) -> Option<&VarEntry> {
+    pub(crate) fn get_by_name(&self, name: &str) -> Option<&VarEntry> {
         self.by_name.get(name).and_then(|b| self.vars.get(b))
     }
 
     /// Mutable resolve by name (cold path).
-    fn get_by_name_mut(&mut self, name: &str) -> Option<&mut VarEntry> {
+    pub(crate) fn get_by_name_mut(&mut self, name: &str) -> Option<&mut VarEntry> {
         let binding = *self.by_name.get(name)?;
         self.vars.get_mut(&binding)
     }
 
     /// Whether a name is declared in this frame.
-    fn contains_name(&self, name: &str) -> bool {
+    pub(crate) fn contains_name(&self, name: &str) -> bool {
         self.by_name.contains_key(name)
     }
 
     /// Insert a fresh variable keyed by its BindingId, with its name indexed.
-    fn insert_var(&mut self, binding: BindingId, name: &str, entry: VarEntry) {
+    pub(crate) fn insert_var(&mut self, binding: BindingId, name: &str, entry: VarEntry) {
         self.vars.insert(binding, entry);
         self.by_name.insert(name.to_string(), binding);
     }
@@ -101,7 +100,7 @@ pub struct RunTime {
     pub handles: HandleRegistry<Value>,
     pub structs: HashMap<String, Vec<(String, Type)>>,
     pub semantic_impls: HashMap<(String, String), (Vec<(BindingId, String, SemanticType)>, Arc<SemanticFunction>)>,
-    scopes: Vec<ScopeFrame>,
+    pub(crate) scopes: Vec<ScopeFrame>,
     pub semantic_funcs: HashMap<String, Arc<SemanticFunction>>,
     pub debug_scope: bool,
     pub consts: HashMap<String, Value>,
@@ -123,17 +122,6 @@ impl RunTime {
         std::str::from_utf8(bytes).expect("arena string was not valid utf8")
     }
 
-    fn resolve_assigned_value(&mut self, value: Value, pos: usize) -> Result<Value, RuntimeError> {
-        match value {
-            Value::Str(off, len) => {
-                let expanded = expand_template(self, self.resolve_str(off, len), pos)?;
-                let (off, len) = self.alloc_str(&expanded);
-                Ok(Value::Str(off, len))
-            }
-            other => Ok(other),
-        }
-    }
-
     pub fn new() -> Self {
         Self {
             string_arena: Vec::new(),
@@ -146,1714 +134,10 @@ impl RunTime {
             consts: HashMap::new(),
         }
     }
-
-    pub fn push_scope(&mut self) {
-        self.scopes.push(ScopeFrame::new());
-        if self.debug_scope {
-            diagnostics::print_scope_event(&ScopeEvent::Open(format!(
-                "scope#{}",
-                self.scopes.len() - 1
-            )));
-        }
-    }
-
-    pub fn push_function_scope(&mut self) {
-        self.scopes.push(ScopeFrame::new());
-        if self.debug_scope {
-            diagnostics::print_scope_event(&ScopeEvent::Open(format!(
-                "scope#{}",
-                self.scopes.len() - 1
-            )));
-        }
-    }
-
-    pub fn pop_scope(&mut self) {
-        if self.scopes.is_empty() {
-            self.scopes.pop();
-            return;
-        }
-
-        let (bleed_values, debug_info) = {
-            let frame = self.scopes.last().unwrap();
-            let bleeds: Vec<(String, usize, String)> = frame
-                .bleed_back
-                .iter()
-                .filter(|(param_name, _)| !frame.freed.contains(*param_name))
-                .map(|(param_name, (outer_idx, outer_name))| {
-                    (param_name.clone(), *outer_idx, outer_name.clone())
-                })
-                .collect();
-
-            let bleed_values: Vec<(usize, String, Value)> = bleeds
-                .iter()
-                .filter_map(|(param_name, outer_idx, outer_name)| {
-                    frame.get_by_name(param_name)
-                        .and_then(|entry| entry.val.clone())
-                        .map(|val| (*outer_idx, outer_name.clone(), val))
-                })
-                .collect();
-
-            let debug_info = if self.debug_scope {
-                let bleed_events: Vec<(String, Value)> = bleeds
-                    .iter()
-                    .filter_map(|(param_name, _, outer_name)| {
-                        frame.get_by_name(param_name)
-                            .and_then(|entry| entry.val.clone())
-                            .map(|val| (outer_name.clone(), val))
-                    })
-                    .collect();
-                let free_names: Vec<String> = frame.by_name.keys()
-                    .filter(|name| !frame.freed.contains(*name))
-                    .cloned()
-                    .collect();
-                let close_label = format!("scope#{}", self.scopes.len() - 1);
-                Some((free_names, bleed_events, close_label))
-            } else {
-                None
-            };
-
-            (bleed_values, debug_info)
-        };
-
-        self.scopes.pop();
-
-        for (outer_idx, outer_name, val) in bleed_values {
-            if let Some(outer_frame) = self.scopes.get_mut(outer_idx) {
-                if let Some(entry) = outer_frame.get_by_name_mut(&outer_name) {
-                    entry.val = Some(val);
-                }
-            }
-        }
-
-        if let Some((free_names, bleed_events, close_label)) = debug_info {
-            for name in &free_names {
-                diagnostics::print_scope_event(&ScopeEvent::Free(name.clone()));
-            }
-            for (name, val) in &bleed_events {
-                diagnostics::print_scope_event(&ScopeEvent::BleedBack(name.clone(), val.clone()));
-            }
-            diagnostics::print_scope_event(&ScopeEvent::Close(close_label));
-        }
-    }
-
-    pub fn declare(
-        &mut self,
-        binding: BindingId,
-        name: String,
-        ty: Option<Type>,
-        pos: usize,
-    ) -> Result<(), RuntimeError> {
-        let frame = self.scopes.last_mut().unwrap();
-
-        if frame.contains_name(&name) {
-            return Err(RuntimeError::AlreadyDeclared { pos, name });
-        }
-
-        frame.seen.insert(name.clone());
-        frame.insert_var(binding, &name, VarEntry { ty, val: None });
-        Ok(())
-    }
-
-    /// Shared write path: type-check `value` against the entry at `binding` in
-    /// frame `i`, write it, emit debug events. The caller has already located the
-    /// owning frame (by binding on the hot path, by name on the cold path).
-    fn write_var_at(
-        &mut self,
-        i: usize,
-        binding: BindingId,
-        name: &str,
-        value: Value,
-        pos: usize,
-    ) -> Result<(), RuntimeError> {
-        let frame = &mut self.scopes[i];
-        let entry = frame.vars.get_mut(&binding).unwrap();
-
-        let was_initialized = entry.val.is_some();
-        if entry.ty.is_none() {
-            entry.ty = Some(type_of_value(&value));
-        }
-
-        let expected = entry.ty.clone().unwrap();
-        let got = type_of_value(&value);
-        if !value_matches_type(&value, &expected) {
-            return Err(RuntimeError::TypeMismatch { pos, expected, got });
-        }
-
-        entry.val = Some(value);
-
-        if self.debug_scope {
-            let logged = entry.val.clone().unwrap();
-            if was_initialized {
-                diagnostics::print_scope_event(&ScopeEvent::Mutate(name.to_string(), logged));
-            } else {
-                diagnostics::print_scope_event(&ScopeEvent::Add(name.to_string(), logged));
-            }
-        }
-        Ok(())
-    }
-
-    /// Hot-path write: assign to the variable identified by `binding`. `name` is
-    /// used only for the const-protection check and diagnostics.
-    pub fn set_var_by_id(
-        &mut self,
-        binding: BindingId,
-        name: &str,
-        value: Value,
-        pos: usize,
-    ) -> Result<(), RuntimeError> {
-        if !self.consts.is_empty() && self.consts.contains_key(name) {
-            return Err(RuntimeError::BadAssignTarget { pos });
-        }
-        let value = self.resolve_assigned_value(value, pos)?;
-        for i in (0..self.scopes.len()).rev() {
-            if self.scopes[i].vars.contains_key(&binding) {
-                return self.write_var_at(i, binding, name, value, pos);
-            }
-        }
-        let was_seen = self.scopes.last().unwrap().seen.contains(name);
-        Err(diagnostics::unresolved_var_error(pos, name.to_string(), was_seen))
-    }
-
-    /// Cold-path write by name (string-interp targets, while-in arrays, method
-    /// write-back). Resolves name -> BindingId via the per-frame index.
-    pub fn set_var(&mut self, name: String, value: Value, pos: usize) -> Result<(), RuntimeError> {
-        if !self.consts.is_empty() && self.consts.contains_key(&name) {
-            return Err(RuntimeError::BadAssignTarget { pos });
-        }
-        let value = self.resolve_assigned_value(value, pos)?;
-        for i in (0..self.scopes.len()).rev() {
-            if let Some(&binding) = self.scopes[i].by_name.get(&name) {
-                return self.write_var_at(i, binding, &name, value, pos);
-            }
-        }
-        let was_seen = self.scopes.last().unwrap().seen.contains(&name);
-        Err(diagnostics::unresolved_var_error(pos, name, was_seen))
-    }
-
-    pub fn set_var_typed(
-        &mut self,
-        binding: BindingId,
-        name: String,
-        ty: Type,
-        value: Value,
-        pos: usize,
-    ) -> Result<(), RuntimeError> {
-        let value = self.resolve_assigned_value(value, pos)?;
-        let logged = value.clone();
-        let got = type_of_value(&value);
-        if !value_matches_type(&value, &ty) {
-            return Err(RuntimeError::TypeMismatch {
-                pos,
-                expected: ty,
-                got,
-            });
-        }
-
-        {
-            let frame = self.scopes.last_mut().unwrap();
-            if frame.contains_name(&name) {
-                return Err(RuntimeError::AlreadyDeclared { pos, name });
-            }
-
-            frame.seen.insert(name.clone());
-            frame.insert_var(
-                binding,
-                &name,
-                VarEntry {
-                    ty: Some(ty),
-                    val: Some(value),
-                },
-            );
-            if self.debug_scope {
-                diagnostics::print_scope_event(&ScopeEvent::Add(name, logged.clone()));
-            }
-        }
-
-        Ok(())
-    }
-
-    pub fn set_container_field(
-        &mut self,
-        container: &str,
-        field: &str,
-        value: Value,
-        pos: usize,
-    ) -> Result<(), RuntimeError> {
-        let logged = value.clone();
-        for frame in self.scopes.iter_mut().rev() {
-            if let Some(entry) = frame.get_by_name_mut(container) {
-                match &mut entry.val {
-                    Some(Value::Container(map)) => {
-                        map.insert(field.to_string(), value);
-                        if self.debug_scope {
-                            diagnostics::print_scope_event(&ScopeEvent::Mutate(
-                                format!("{}.{}", container, field),
-                                logged,
-                            ));
-                        }
-                        return Ok(());
-                    }
-                    Some(Value::Struct(_, map)) => {
-                        map.insert(field.to_string(), value);
-                        if self.debug_scope {
-                            diagnostics::print_scope_event(&ScopeEvent::Mutate(
-                                format!("{}.{}", container, field),
-                                logged,
-                            ));
-                        }
-                        return Ok(());
-                    }
-                    _ => {
-                        return Err(RuntimeError::NotAContainer {
-                            pos,
-                            name: container.to_string(),
-                        });
-                    }
-                }
-            }
-        }
-        Err(RuntimeError::UndefinedVar {
-            pos,
-            name: container.to_string(),
-        })
-    }
-
-    /// Write a single element of an array variable in scope.
-    ///
-    /// Traverses scopes from innermost outward, finds the named variable,
-    /// and replaces the element at `index` with `value`.  Returns an error
-    /// if the variable is not found, is not an array, or the index is
-    /// out of bounds.
-    pub fn set_array_element(
-        &mut self,
-        arr_name: &str,
-        index: usize,
-        value: Value,
-        pos: usize,
-    ) -> Result<(), RuntimeError> {
-        for frame in self.scopes.iter_mut().rev() {
-            if let Some(entry) = frame.get_by_name_mut(arr_name) {
-                match &mut entry.val {
-                    Some(Value::Array(elems)) => {
-                        if index < elems.len() {
-                            elems[index] = value;
-                            return Ok(());
-                        } else {
-                            return Err(RuntimeError::IndexOutOfBounds {
-                                pos,
-                                index: index as i64,
-                                length: elems.len(),
-                            });
-                        }
-                    }
-                    _ => {
-                        return Err(RuntimeError::NotAContainer {
-                            pos,
-                            name: arr_name.to_string(),
-                        });
-                    }
-                }
-            }
-        }
-        Err(RuntimeError::UndefinedVar {
-            pos,
-            name: arr_name.to_string(),
-        })
-    }
-
-    pub fn get_field(&self, container: &str, field: &str, pos: usize) -> Result<Value, RuntimeError> {
-        for frame in self.scopes.iter().rev() {
-            if let Some(entry) = frame.get_by_name(container) {
-                match &entry.val {
-                    Some(Value::Struct(_, map)) | Some(Value::Container(map)) => {
-                        return map.get(field).cloned().ok_or_else(|| {
-                            RuntimeError::UndefinedVar { pos, name: format!("{}.{}", container, field) }
-                        });
-                    }
-                    _ => return Err(RuntimeError::NotAContainer { pos, name: container.to_string() }),
-                }
-            }
-        }
-        Err(RuntimeError::UndefinedVar { pos, name: container.to_string() })
-    }
-
-    /// Cold-path read by name (string interpolation, container/array name
-    /// resolution, `.copy` args). Resolves name -> BindingId via the per-frame
-    /// index, then reads.
-    pub fn get_var(&self, name: &str, pos: usize) -> Result<Value, RuntimeError> {
-        for frame in self.scopes.iter().rev() {
-            if let Some(entry) = frame.get_by_name(name) {
-                if let Some(value) = &entry.val {
-                    return Ok(value.clone());
-                }
-                return Err(RuntimeError::UninitializedVar {
-                    pos,
-                    name: name.to_string(),
-                });
-            }
-        }
-        let owned = name.to_string();
-        let was_seen = self.scopes.last().unwrap().seen.contains(&owned);
-        Err(diagnostics::unresolved_var_error(pos, owned, was_seen))
-    }
-
-    /// Hot-path read by BindingId (VarRef, CompoundAssign read). `name` is used
-    /// only for diagnostics. Walks frames innermost->outermost exactly like the
-    /// name-based path, but matches on the BindingId — no string hashing.
-    pub fn get_var_by_id(&self, binding: BindingId, name: &str, pos: usize) -> Result<Value, RuntimeError> {
-        for frame in self.scopes.iter().rev() {
-            if let Some(entry) = frame.vars.get(&binding) {
-                if let Some(value) = &entry.val {
-                    return Ok(value.clone());
-                }
-                return Err(RuntimeError::UninitializedVar {
-                    pos,
-                    name: name.to_string(),
-                });
-            }
-        }
-        let was_seen = self.scopes.last().unwrap().seen.contains(name);
-        Err(diagnostics::unresolved_var_error(pos, name.to_string(), was_seen))
-    }
-
-    fn apply_op(
-        &self,
-        left: Value,
-        op: Op,
-        pos: usize,
-        right: Value,
-    ) -> Result<Value, RuntimeError> {
-        match (&left, &op, &right) {
-            (Value::Bool(false), Op::And, Value::Unknown(_)) => return Ok(Value::Bool(false)),
-            (Value::Unknown(_), Op::And, Value::Bool(false)) => return Ok(Value::Bool(false)),
-            (Value::Bool(true), Op::Or, Value::Unknown(_)) => return Ok(Value::Bool(true)),
-            (Value::Unknown(_), Op::Or, Value::Bool(true)) => return Ok(Value::Bool(true)),
-            (Value::TBool(0), Op::And, _) => return Ok(Value::TBool(0)),
-            (_, Op::And, Value::TBool(0)) => return Ok(Value::TBool(0)),
-            (Value::TBool(1), Op::Or, _) => return Ok(Value::TBool(1)),
-            (_, Op::Or, Value::TBool(1)) => return Ok(Value::TBool(1)),
-            (Value::Unknown(_), Op::Mul, Value::Num(0)) => return Ok(Value::Num(0)),
-            (Value::Num(0), Op::Mul, Value::Unknown(_)) => return Ok(Value::Num(0)),
-            (Value::Unknown(ty), _, _) => return Ok(Value::Unknown(ty.clone())),
-            (_, _, Value::Unknown(ty)) => return Ok(Value::Unknown(ty.clone())),
-            _ => {}
-        }
-
-        match op {
-            Op::Plus => match (&left, &right) {
-                (Value::Num(a), Value::Num(b)) => Ok(Value::Num(a.wrapping_add(*b))),
-                _ => {
-                    if let (Some(a), Some(b)) = (as_f64(&left), as_f64(&right)) {
-                        Ok(Value::Float(a + b))
-                    } else {
-                        Err(RuntimeError::BadOperands {
-                            pos,
-                            op,
-                            left,
-                            right,
-                        })
-                    }
-                }
-            },
-            Op::Minus => match (&left, &right) {
-                (Value::Num(a), Value::Num(b)) => Ok(Value::Num(a.wrapping_sub(*b))),
-                _ => {
-                    if let (Some(a), Some(b)) = (as_f64(&left), as_f64(&right)) {
-                        Ok(Value::Float(a - b))
-                    } else {
-                        Err(RuntimeError::BadOperands {
-                            pos,
-                            op,
-                            left,
-                            right,
-                        })
-                    }
-                }
-            },
-            Op::Mul => match (&left, &right) {
-                (Value::Num(a), Value::Num(b)) => {
-                    Ok(Value::Num(a.wrapping_mul(*b)))
-                }
-                _ => {
-                    if let (Some(a), Some(b)) = (as_f64(&left), as_f64(&right)) {
-                        Ok(Value::Float(a * b))
-                    } else {
-                        Err(RuntimeError::BadOperands {
-                            pos,
-                            op,
-                            left,
-                            right,
-                        })
-                    }
-                }
-            },
-            Op::Div => match (&left, &right) {
-                (Value::Num(_), Value::Num(0)) => Err(RuntimeError::DivByZero { pos }),
-                (Value::Num(a), Value::Num(b)) => Ok(Value::Num(if *b == -1 { a.wrapping_neg() } else { a / b })),
-                _ => {
-                    if let (Some(a), Some(b)) = (as_f64(&left), as_f64(&right)) {
-                        if b == 0.0 {
-                            Err(RuntimeError::DivByZero { pos })
-                        } else {
-                            Ok(Value::Float(a / b))
-                        }
-                    } else {
-                        Err(RuntimeError::BadOperands {
-                            pos,
-                            op,
-                            left,
-                            right,
-                        })
-                    }
-                }
-            },
-            Op::Mod => match (&left, &right) {
-                (Value::Num(_), Value::Num(0)) => Err(RuntimeError::DivByZero { pos }),
-                (Value::Num(a), Value::Num(b)) => Ok(Value::Num(if *b == -1 { 0 } else { a % b })),
-                _ => {
-                    if let (Some(a), Some(b)) = (as_f64(&left), as_f64(&right)) {
-                        if b == 0.0 {
-                            Err(RuntimeError::DivByZero { pos })
-                        } else {
-                            Ok(Value::Float(a % b))
-                        }
-                    } else {
-                        Err(RuntimeError::BadOperands {
-                            pos,
-                            op,
-                            left,
-                            right,
-                        })
-                    }
-                }
-            },
-            Op::EqEq => match (&left, &right) {
-                (Value::Num(a), Value::Num(b)) => Ok(Value::Bool(a == b)),
-                (Value::Float(a), Value::Float(b)) => Ok(Value::Bool(a == b)),
-                (Value::Num(a), Value::Float(b)) => Ok(Value::Bool((*a as f64) == *b)),
-                (Value::Float(a), Value::Num(b)) => Ok(Value::Bool(*a == (*b as f64))),
-                (Value::Str(a_off, a_len), Value::Str(b_off, b_len)) => Ok(Value::Bool(
-                    self.resolve_str(*a_off, *a_len) == self.resolve_str(*b_off, *b_len),
-                )),
-                (Value::Char(a), Value::Char(b)) => Ok(Value::Bool(a == b)),
-                (Value::Bool(a), Value::Bool(b)) => Ok(Value::Bool(a == b)),
-                (Value::TBool(2), _) | (_, Value::TBool(2)) => Ok(Value::TBool(2)),
-                (Value::TBool(a), Value::TBool(b)) => Ok(Value::Bool(a == b)),
-                (Value::TBool(a), Value::Bool(b)) => Ok(Value::Bool((*a == 1) == *b)),
-                (Value::Bool(a), Value::TBool(b)) => Ok(Value::Bool(*a == (*b == 1))),
-                (l, r) => Err(RuntimeError::BadOperands {
-                    pos,
-                    op,
-                    left: l.clone(),
-                    right: r.clone(),
-                }),
-            },
-            Op::NotEq => match (&left, &right) {
-                (Value::Num(a), Value::Num(b)) => Ok(Value::Bool(a != b)),
-                (Value::Float(a), Value::Float(b)) => Ok(Value::Bool(a != b)),
-                (Value::Num(a), Value::Float(b)) => Ok(Value::Bool((*a as f64) != *b)),
-                (Value::Float(a), Value::Num(b)) => Ok(Value::Bool(*a != (*b as f64))),
-                (Value::Str(a_off, a_len), Value::Str(b_off, b_len)) => Ok(Value::Bool(
-                    self.resolve_str(*a_off, *a_len) != self.resolve_str(*b_off, *b_len),
-                )),
-                (Value::Char(a), Value::Char(b)) => Ok(Value::Bool(a != b)),
-                (Value::Bool(a), Value::Bool(b)) => Ok(Value::Bool(a != b)),
-                (Value::TBool(2), _) | (_, Value::TBool(2)) => Ok(Value::TBool(2)),
-                (Value::TBool(a), Value::TBool(b)) => Ok(Value::Bool(a != b)),
-                (Value::TBool(a), Value::Bool(b)) => Ok(Value::Bool((*a == 1) != *b)),
-                (Value::Bool(a), Value::TBool(b)) => Ok(Value::Bool(*a != (*b == 1))),
-                (l, r) => Err(RuntimeError::BadOperands {
-                    pos,
-                    op,
-                    left: l.clone(),
-                    right: r.clone(),
-                }),
-            },
-            Op::Lt => match (&left, &right) {
-                (Value::Num(a), Value::Num(b)) => Ok(Value::Bool(a < b)),
-                _ => Err(RuntimeError::BadOperands {
-                    pos,
-                    op,
-                    left,
-                    right,
-                }),
-            },
-            Op::Gt => match (&left, &right) {
-                (Value::Num(a), Value::Num(b)) => Ok(Value::Bool(a > b)),
-                _ => Err(RuntimeError::BadOperands {
-                    pos,
-                    op,
-                    left,
-                    right,
-                }),
-            },
-            Op::LtEq => match (&left, &right) {
-                (Value::Num(a), Value::Num(b)) => Ok(Value::Bool(a <= b)),
-                _ => Err(RuntimeError::BadOperands {
-                    pos,
-                    op,
-                    left,
-                    right,
-                }),
-            },
-            Op::GtEq => match (&left, &right) {
-                (Value::Num(a), Value::Num(b)) => Ok(Value::Bool(a >= b)),
-                _ => Err(RuntimeError::BadOperands {
-                    pos,
-                    op,
-                    left,
-                    right,
-                }),
-            },
-            Op::Not => unreachable!("Op::Not is unary only"),
-            Op::And => match (&left, &right) {
-                (Value::Bool(a), Value::Bool(b)) => Ok(Value::Bool(*a && *b)),
-                (Value::TBool(a), Value::TBool(b)) => Ok(Value::TBool(match (a, b) {
-                    (0, _) | (_, 0) => 0,
-                    (1, 1) => 1,
-                    _ => 2,
-                })),
-                (Value::Bool(true), Value::TBool(b)) => Ok(Value::TBool(*b)),
-                (Value::TBool(a), Value::Bool(true)) => Ok(Value::TBool(*a)),
-                (Value::Bool(false), Value::TBool(_)) => Ok(Value::TBool(0)),
-                (Value::TBool(_), Value::Bool(false)) => Ok(Value::TBool(0)),
-                (l, r) => Err(RuntimeError::BadOperands {
-                    pos,
-                    op,
-                    left: l.clone(),
-                    right: r.clone(),
-                }),
-            },
-            Op::Or => match (&left, &right) {
-                (Value::Bool(a), Value::Bool(b)) => Ok(Value::Bool(*a || *b)),
-                (Value::TBool(a), Value::TBool(b)) => Ok(Value::TBool(match (a, b) {
-                    (1, _) | (_, 1) => 1,
-                    (0, 0) => 0,
-                    _ => 2,
-                })),
-                (Value::Bool(false), Value::TBool(b)) => Ok(Value::TBool(*b)),
-                (Value::TBool(a), Value::Bool(false)) => Ok(Value::TBool(*a)),
-                (Value::Bool(true), Value::TBool(_)) => Ok(Value::TBool(1)),
-                (Value::TBool(_), Value::Bool(true)) => Ok(Value::TBool(1)),
-                (l, r) => Err(RuntimeError::BadOperands {
-                    pos,
-                    op,
-                    left: l.clone(),
-                    right: r.clone(),
-                }),
-            },
-        }
-    }
-
-    // ── Semantic IR interpreter ──────────────────────────────────────
-
-    pub fn eval_semantic_expr(&mut self, expr: &SemanticExpr) -> Result<Value, RuntimeError> {
-        match &expr.kind {
-            SemanticExprKind::Value(sv) => Ok(self.semantic_value_to_runtime(sv)),
-            SemanticExprKind::VarRef { binding, name } => {
-                self.get_var_by_id(*binding, name, 0)
-            }
-            SemanticExprKind::Unary { op, expr: inner, pos } => {
-                let val = self.eval_semantic_expr(inner)?;
-                let result = self.apply_unary(op, val, *pos)?;
-                Ok(apply_numeric_cast(result, &expr.ty))
-            }
-            SemanticExprKind::Binary { lhs, op, pos, rhs } => {
-                // Short-circuit semantics for && and ||: evaluate lhs first; if
-                // the result is already determined, skip rhs entirely.  This
-                // matches the CFG-based JIT lowering added in CX-105.
-                if matches!(op, Op::And | Op::Or) {
-                    let l = self.eval_semantic_expr(lhs)?;
-                    let skip = match (&l, op) {
-                        (Value::Bool(false), Op::And) | (Value::TBool(0), Op::And) => true,
-                        (Value::Bool(true), Op::Or)  | (Value::TBool(1), Op::Or)  => true,
-                        _ => false,
-                    };
-                    if skip {
-                        return Ok(apply_numeric_cast(l, &expr.ty));
-                    }
-                    let r = self.eval_semantic_expr(rhs)?;
-                    let result = self.apply_op(l, op.clone(), *pos, r)?;
-                    return Ok(apply_numeric_cast(result, &expr.ty));
-                }
-                // Evaluation order guarantee: lhs is fully evaluated before rhs.
-                // Any side effects in lhs (e.g. function calls with print) occur
-                // before any side effects in rhs.  The IR lowering (lower_binary
-                // in ir/lower.rs) mirrors this order exactly.  This is a Cx 0.1
-                // language guarantee.  See docs/backend/cx_eval_order.md.
-                let l = self.eval_semantic_expr(lhs)?;
-                let r = self.eval_semantic_expr(rhs)?;
-                let result = self.apply_op(l, op.clone(), *pos, r)?;
-                Ok(apply_numeric_cast(result, &expr.ty))
-            }
-            SemanticExprKind::Call { callee, args, .. } => {
-                self.call_semantic_func(callee, args, 0)
-            }
-            SemanticExprKind::DotAccess { container, field, .. } => {
-                self.get_field(container, field, 0)
-            }
-            SemanticExprKind::StructInstance { type_name, fields } => {
-                let mut map = HashMap::new();
-                for (fname, fexpr) in fields {
-                    let val = self.eval_semantic_expr(fexpr)?;
-                    map.insert(fname.clone(), val);
-                }
-                Ok(Value::Struct(type_name.clone(), map))
-            }
-            SemanticExprKind::ArrayLit { elements } => {
-                let mut vals = Vec::new();
-                for e in elements {
-                    vals.push(self.eval_semantic_expr(e)?);
-                }
-                Ok(Value::Array(vals))
-            }
-            SemanticExprKind::Index { target, index, pos } => {
-                let arr = self.eval_semantic_expr(target)?;
-                let idx = self.eval_semantic_expr(index)?;
-                match (arr, idx) {
-                    (Value::Array(elems), Value::Num(i)) => {
-                        let length = elems.len();
-                        elems.get(i as usize).cloned().ok_or(RuntimeError::IndexOutOfBounds { pos: *pos, index: i as i64, length })
-                    }
-                    _ => Err(RuntimeError::BadAssignTarget { pos: *pos })
-                }
-            }
-            SemanticExprKind::When { expr, arms, .. } => {
-                let val = self.eval_semantic_expr(expr)?;
-                let result = self.run_semantic_when(val, arms)?;
-                Ok(result)
-            }
-            SemanticExprKind::MethodCall { instance, method, args, pos, .. } => {
-                self.call_semantic_method(instance, method, args, *pos)
-            }
-            SemanticExprKind::Range { .. } => Ok(Value::Num(0)), // stub
-            SemanticExprKind::HandleNew { value, .. } => {
-                let val = self.eval_semantic_expr(value)?;
-                let h = self.handles.insert(val);
-                Ok(Value::Handle(h))
-            }
-            SemanticExprKind::HandleVal { name, pos, .. } => {
-                let val = self.get_var(name, *pos)?;
-                if let Value::Handle(h) = val {
-                    match self.handles.get(h) {
-                        Some(v) => Ok(v.clone()),
-                        None => Err(RuntimeError::StaleHandle { pos: *pos }),
-                    }
-                } else {
-                    Err(RuntimeError::StaleHandle { pos: *pos })
-                }
-            }
-            SemanticExprKind::HandleDrop { name, pos, .. } => {
-                let val = self.get_var(name, *pos)?;
-                if let Value::Handle(h) = val {
-                    self.handles.remove(h);
-                    Ok(Value::Num(0))
-                } else {
-                    Err(RuntimeError::StaleHandle { pos: *pos })
-                }
-            }
-            SemanticExprKind::ResultOk { expr } => {
-                let val = self.eval_semantic_expr(expr)?;
-                Ok(Value::ResultOk(Box::new(val)))
-            }
-            SemanticExprKind::ResultErr { expr } => {
-                let val = self.eval_semantic_expr(expr)?;
-                Ok(Value::ResultErr(Box::new(val)))
-            }
-            SemanticExprKind::Try { expr, .. } => {
-                let val = self.eval_semantic_expr(expr)?;
-                match val {
-                    Value::ResultOk(v) => Ok(*v),
-                    Value::ResultErr(e) => Err(RuntimeError::EarlyReturn(Value::ResultErr(e))),
-                    _ => Ok(val),
-                }
-            }
-            SemanticExprKind::Cast { expr, to, .. } => {
-                let val = self.eval_semantic_expr(expr)?;
-                Ok(apply_numeric_cast(val, to))
-            }
-        }
-    }
-
-    pub fn run_semantic_stmt(&mut self, stmt: &SemanticStmt) -> Result<(), RuntimeError> {
-        match stmt {
-            SemanticStmt::Noop => Ok(()),
-SemanticStmt::Decl { binding, name, ty, .. } => {
-                let rt_ty: Option<Type> = ty.as_ref().map(|t| t.clone().into());
-                self.declare(*binding, name.clone(), rt_ty, 0)
-            }
-            SemanticStmt::TypedAssign { binding, name, ty, expr, .. } => {
-                let val = self.eval_semantic_expr(expr)?;
-                let rt_ty: Type = ty.clone().into();
-                let val = match (&rt_ty, val) {
-                    (Type::Bool, Value::Unknown(_)) => Value::TBool(2),
-                    (_, v) => v,
-                };
-                self.set_var_typed(*binding, name.clone(), rt_ty, val, 0)
-            }
-            SemanticStmt::Assign { target, expr, pos_eq } => {
-                let val = self.eval_semantic_expr(expr)?;
-                match target {
-                    SemanticLValue::Binding { binding, name, ty } => {
-                        let truncated = apply_numeric_cast(val, ty);
-                        self.set_var_by_id(*binding, name, truncated, 0)
-                    }
-                    SemanticLValue::DotAccess { container, field, ty, .. } => {
-                        let truncated = apply_numeric_cast(val, ty);
-                        self.set_container_field(container, field, truncated, 0)
-                    }
-                    SemanticLValue::Index { target, index, elem_ty } => {
-                        let arr_name = match &target.kind {
-                            SemanticExprKind::VarRef { name, .. } => name.clone(),
-                            _ => return Err(RuntimeError::BadAssignTarget { pos: 0 }),
-                        };
-                        let idx = match self.eval_semantic_expr(index)? {
-                            Value::Num(n) => n as usize,
-                            _ => return Err(RuntimeError::BadAssignTarget { pos: 0 }),
-                        };
-                        let truncated = apply_numeric_cast(val, elem_ty);
-                        self.set_array_element(&arr_name, idx, truncated, *pos_eq)
-                    }
-                }
-            }
-            SemanticStmt::CompoundAssign { target, op, operand, pos } => {
-                match target {
-                    SemanticLValue::Binding { binding, name, ty } => {
-                        let current = self.get_var_by_id(*binding, name, 0)?;
-                        let rhs = self.eval_semantic_expr(operand)?;
-                        let result = self.apply_op(current, op.clone(), 0, rhs)?;
-                        let truncated = apply_numeric_cast(result, ty);
-                        self.set_var_by_id(*binding, name, truncated, 0)
-                    }
-                    SemanticLValue::DotAccess { container, field, ty, .. } => {
-                        let current = self.get_field(container, field, 0)?;
-                        let rhs = self.eval_semantic_expr(operand)?;
-                        let result = self.apply_op(current, op.clone(), 0, rhs)?;
-                        let truncated = apply_numeric_cast(result, ty);
-                        self.set_container_field(container, field, truncated, 0)
-                    }
-                    SemanticLValue::Index { target, index, elem_ty } => {
-                        let arr_name = match &target.kind {
-                            SemanticExprKind::VarRef { name, .. } => name.clone(),
-                            _ => return Err(RuntimeError::BadAssignTarget { pos: 0 }),
-                        };
-                        let idx = match self.eval_semantic_expr(index)? {
-                            Value::Num(n) => n as usize,
-                            _ => return Err(RuntimeError::BadAssignTarget { pos: 0 }),
-                        };
-                        let current_val = {
-                            let arr = self.get_var(&arr_name, 0)?;
-                            match arr {
-                                Value::Array(elems) => {
-                                    let length = elems.len();
-                                    elems
-                                        .get(idx)
-                                        .cloned()
-                                        .ok_or(RuntimeError::IndexOutOfBounds {
-                                            pos: *pos,
-                                            index: idx as i64,
-                                            length,
-                                        })?
-                                }
-                                _ => return Err(RuntimeError::NotAContainer {
-                                    pos: 0,
-                                    name: arr_name.clone(),
-                                }),
-                            }
-                        };
-                        let rhs = self.eval_semantic_expr(operand)?;
-                        let result = self.apply_op(current_val, op.clone(), 0, rhs)?;
-                        let truncated = apply_numeric_cast(result, elem_ty);
-                        self.set_array_element(&arr_name, idx, truncated, *pos)
-                    }
-                }
-            }
-            SemanticStmt::Return { expr, .. } => {
-                if let Some(e) = expr {
-                    let val = self.eval_semantic_expr(e)?;
-                    Err(RuntimeError::EarlyReturn(val))
-                } else {
-                    Err(RuntimeError::EarlyReturn(Value::Num(0)))
-                }
-            }
-            SemanticStmt::ExprStmt { expr, .. } => {
-                self.eval_semantic_expr(expr)?;
-                Ok(())
-            }
-            SemanticStmt::Block { stmts, .. } => {
-                self.push_scope();
-                for s in stmts {
-                    match self.run_semantic_stmt(s) {
-                        Ok(_) => {}
-                        Err(e) => { self.pop_scope(); return Err(e); }
-                    }
-                }
-                self.pop_scope();
-                Ok(())
-            }
-            SemanticStmt::While { cond, body, .. } => {
-                loop {
-                    let cv = self.eval_semantic_expr(cond)?;
-                    match cv {
-                        Value::Bool(false) | Value::TBool(0) => break,
-                        Value::Bool(true) | Value::TBool(1) => {}
-                        _ => break,
-                    }
-                    self.push_scope();
-                    let mut should_break = false;
-                    for s in body {
-                        match self.run_semantic_stmt(s) {
-                            Ok(_) => {}
-                            Err(RuntimeError::BreakSignal) => { should_break = true; break; }
-                            Err(RuntimeError::ContinueSignal) => break,
-                            Err(e) => { self.pop_scope(); return Err(e); }
-                        }
-                    }
-                    self.pop_scope();
-                    if should_break { break; }
-                }
-                Ok(())
-            }
-            SemanticStmt::For { binding, var, start, end, inclusive, body, .. } => {
-                let start_val = match self.eval_semantic_expr(start)? {
-                    Value::Num(n) => n,
-                    _ => return Err(RuntimeError::BadAssignTarget { pos: 0 }),
-                };
-                let end_val = match self.eval_semantic_expr(end)? {
-                    Value::Num(n) => n,
-                    _ => return Err(RuntimeError::BadAssignTarget { pos: 0 }),
-                };
-                'sem_for: {
-                    if *inclusive {
-                        for i in start_val..=end_val {
-                            self.push_scope();
-                            self.declare(*binding, var.clone(), None, 0)?;
-                            self.set_var_by_id(*binding, var, Value::Num(i), 0)?;
-                            for s in body {
-                                match self.run_semantic_stmt(s) {
-                                    Ok(_) => {}
-                                    Err(RuntimeError::BreakSignal) => { self.pop_scope(); break 'sem_for; }
-                                    Err(RuntimeError::ContinueSignal) => break,
-                                    Err(e) => { self.pop_scope(); return Err(e); }
-                                }
-                            }
-                            self.pop_scope();
-                        }
-                    } else {
-                        for i in start_val..end_val {
-                            self.push_scope();
-                            self.declare(*binding, var.clone(), None, 0)?;
-                            self.set_var_by_id(*binding, var, Value::Num(i), 0)?;
-                            for s in body {
-                                match self.run_semantic_stmt(s) {
-                                    Ok(_) => {}
-                                    Err(RuntimeError::BreakSignal) => { self.pop_scope(); break 'sem_for; }
-                                    Err(RuntimeError::ContinueSignal) => break,
-                                    Err(e) => { self.pop_scope(); return Err(e); }
-                                }
-                            }
-                            self.pop_scope();
-                        }
-                    }
-                }
-                Ok(())
-            }
-            SemanticStmt::Loop { body, .. } => {
-                loop {
-                    self.push_scope();
-                    let mut should_break = false;
-                    for s in body {
-                        match self.run_semantic_stmt(s) {
-                            Ok(_) => {}
-                            Err(RuntimeError::BreakSignal) => { should_break = true; break; }
-                            Err(RuntimeError::ContinueSignal) => break,
-                            Err(e) => { self.pop_scope(); return Err(e); }
-                        }
-                    }
-                    self.pop_scope();
-                    if should_break { break; }
-                }
-                Ok(())
-            }
-            SemanticStmt::Break { .. } => Err(RuntimeError::BreakSignal),
-            SemanticStmt::Continue { .. } => Err(RuntimeError::ContinueSignal),
-            SemanticStmt::FuncDef(sem_func) => {
-                self.semantic_funcs.insert(sem_func.name.clone(), Arc::new(sem_func.clone()));
-                Ok(())
-            }
-            SemanticStmt::StructDef { name, fields, .. } => {
-                self.structs.insert(name.clone(), fields.iter().map(|(n, t)| (n.clone(), t.clone().into())).collect());
-                Ok(())
-            }
-            SemanticStmt::ImplBlock { aliases, methods, method_alias_params, .. } => {
-                for (mi, sem_func) in methods.iter().enumerate() {
-                    // Carry the per-method alias BindingIds (the ids the method
-                    // body's alias VarRefs were resolved against, #009) alongside
-                    // each alias name+type. method_alias_params[mi] is parallel to
-                    // `aliases` (both in declaration order).
-                    let method_aliases: Vec<(BindingId, String, SemanticType)> = method_alias_params
-                        .get(mi)
-                        .map(|params| {
-                            params
-                                .iter()
-                                .enumerate()
-                                .map(|(ai, p)| {
-                                    let ty = aliases
-                                        .get(ai)
-                                        .map(|(_, t)| t.clone())
-                                        .or_else(|| p.ty.clone())
-                                        .unwrap_or(SemanticType::Unknown);
-                                    (p.binding, p.name.clone(), ty)
-                                })
-                                .collect()
-                        })
-                        .unwrap_or_default();
-                    for (_, alias_type) in aliases {
-                        let type_key = match alias_type {
-                            SemanticType::Struct(n) => n.clone(),
-                            _ => continue,
-                        };
-                        self.semantic_impls.insert(
-                            (type_key, sem_func.name.clone()),
-                            (method_aliases.clone(), Arc::new(sem_func.clone())),
-                        );
-                    }
-                }
-                Ok(())
-            }
-            SemanticStmt::ConstDecl { binding, name, ty, value, .. } => {
-                let val = self.eval_semantic_expr(value)?;
-                self.consts.insert(name.clone(), val.clone());
-                let ast_ty: Type = ty.clone().into();
-                self.set_var_typed(*binding, name.clone(), ast_ty, val, 0)?;
-                Ok(())
-            }
-            SemanticStmt::EnumDef { .. } => {
-                // Enum variants are resolved at semantic analysis time
-                // Runtime registration not needed — when matching uses SemanticWhenPattern
-                Ok(())
-            }
-            SemanticStmt::When { expr, arms, .. } => {
-                let val = self.eval_semantic_expr(expr)?;
-                self.run_semantic_when(val, arms)?;
-                Ok(())
-            }
-            SemanticStmt::IfElse { condition, then_body, else_ifs, else_body, .. } => {
-                let cond_val = self.eval_semantic_expr(condition)?;
-                let is_true = match &cond_val {
-                    Value::Bool(b) => *b,
-                    Value::TBool(0) => false,
-                    Value::TBool(1) => true,
-                    Value::Num(n) => *n != 0,
-                    _ => false,
-                };
-
-                if is_true {
-                    self.push_scope();
-                    for stmt in then_body {
-                        match self.run_semantic_stmt(stmt) {
-                            Ok(_) => {}
-                            Err(e) => { self.pop_scope(); return Err(e); }
-                        }
-                    }
-                    self.pop_scope();
-                    return Ok(());
-                }
-
-                for (else_if_cond, else_if_body) in else_ifs {
-                    let cond_val = self.eval_semantic_expr(else_if_cond)?;
-                    let is_true = match &cond_val {
-                        Value::Bool(b) => *b,
-                        Value::TBool(0) => false,
-                        Value::TBool(1) => true,
-                        Value::Num(n) => *n != 0,
-                        _ => false,
-                    };
-                    if is_true {
-                        self.push_scope();
-                        for stmt in else_if_body {
-                            match self.run_semantic_stmt(stmt) {
-                                Ok(_) => {}
-                                Err(e) => { self.pop_scope(); return Err(e); }
-                            }
-                        }
-                        self.pop_scope();
-                        return Ok(());
-                    }
-                }
-
-                if let Some(else_body) = else_body {
-                    self.push_scope();
-                    for stmt in else_body {
-                        match self.run_semantic_stmt(stmt) {
-                            Ok(_) => {}
-                            Err(e) => { self.pop_scope(); return Err(e); }
-                        }
-                    }
-                    self.pop_scope();
-                }
-
-                Ok(())
-            }
-            SemanticStmt::WhileIn { arr, start_slot, range_start, range_end, inclusive, body, then_chains, result, .. } => {
-                let start_val = match self.eval_semantic_expr(range_start)? {
-                    Value::Num(n) => n,
-                    _ => return Err(RuntimeError::BadAssignTarget { pos: 0 }),
-                };
-                let end_val = match self.eval_semantic_expr(range_end)? {
-                    Value::Num(n) => n,
-                    _ => return Err(RuntimeError::BadAssignTarget { pos: 0 }),
-                };
-
-                // Primary chain
-                'primary: {
-                    if *inclusive {
-                        for i in start_val..=end_val {
-                            let elem = {
-                                let arr_val = self.get_var(arr, 0)?;
-                                match arr_val {
-                                    Value::Array(elems) => elems.get(i as usize).cloned().unwrap_or(Value::Unknown(Type::Unknown)),
-                                    _ => return Err(RuntimeError::NotAContainer { pos: 0, name: arr.clone() }),
-                                }
-                            };
-                            {
-                                let arr_val = self.get_var(arr, 0)?;
-                                if let Value::Array(mut elems) = arr_val {
-                                    if *start_slot < elems.len() { elems[*start_slot] = elem; }
-                                    self.set_var(arr.clone(), Value::Array(elems), 0)?;
-                                }
-                            }
-                            self.push_scope();
-                            let mut should_break = false;
-                            for stmt in body {
-                                match self.run_semantic_stmt(stmt) {
-                                    Ok(_) => {}
-                                    Err(RuntimeError::BreakSignal) => { should_break = true; break; }
-                                    Err(RuntimeError::ContinueSignal) => break,
-                                    Err(e) => { self.pop_scope(); return Err(e); }
-                                }
-                            }
-                            self.pop_scope();
-                            if should_break { break 'primary; }
-                        }
-                    } else {
-                        for i in start_val..end_val {
-                            let elem = {
-                                let arr_val = self.get_var(arr, 0)?;
-                                match arr_val {
-                                    Value::Array(elems) => elems.get(i as usize).cloned().unwrap_or(Value::Unknown(Type::Unknown)),
-                                    _ => return Err(RuntimeError::NotAContainer { pos: 0, name: arr.clone() }),
-                                }
-                            };
-                            {
-                                let arr_val = self.get_var(arr, 0)?;
-                                if let Value::Array(mut elems) = arr_val {
-                                    if *start_slot < elems.len() { elems[*start_slot] = elem; }
-                                    self.set_var(arr.clone(), Value::Array(elems), 0)?;
-                                }
-                            }
-                            self.push_scope();
-                            let mut should_break = false;
-                            for stmt in body {
-                                match self.run_semantic_stmt(stmt) {
-                                    Ok(_) => {}
-                                    Err(RuntimeError::BreakSignal) => { should_break = true; break; }
-                                    Err(RuntimeError::ContinueSignal) => break,
-                                    Err(e) => { self.pop_scope(); return Err(e); }
-                                }
-                            }
-                            self.pop_scope();
-                            if should_break { break 'primary; }
-                        }
-                    }
-                }
-
-                // Then chains
-                for chain in then_chains {
-                    let chain_start = match self.eval_semantic_expr(&chain.range_start)? {
-                        Value::Num(n) => n,
-                        _ => return Err(RuntimeError::BadAssignTarget { pos: 0 }),
-                    };
-                    let chain_end = match self.eval_semantic_expr(&chain.range_end)? {
-                        Value::Num(n) => n,
-                        _ => return Err(RuntimeError::BadAssignTarget { pos: 0 }),
-                    };
-                    if chain.inclusive {
-                        for i in chain_start..=chain_end {
-                            let elem = {
-                                let arr_val = self.get_var(&chain.arr, 0)?;
-                                match arr_val {
-                                    Value::Array(elems) => elems.get(i as usize).cloned().unwrap_or(Value::Unknown(Type::Unknown)),
-                                    _ => return Err(RuntimeError::NotAContainer { pos: 0, name: chain.arr.clone() }),
-                                }
-                            };
-                            {
-                                let arr_val = self.get_var(&chain.arr, 0)?;
-                                if let Value::Array(mut elems) = arr_val {
-                                    if chain.start_slot < elems.len() { elems[chain.start_slot] = elem; }
-                                    self.set_var(chain.arr.clone(), Value::Array(elems), 0)?;
-                                }
-                            }
-                            self.push_scope();
-                            let mut should_break = false;
-                            for stmt in &chain.body {
-                                match self.run_semantic_stmt(stmt) {
-                                    Ok(_) => {}
-                                    Err(RuntimeError::BreakSignal) => { should_break = true; break; }
-                                    Err(RuntimeError::ContinueSignal) => break,
-                                    Err(e) => { self.pop_scope(); return Err(e); }
-                                }
-                            }
-                            self.pop_scope();
-                            if should_break { break; }
-                        }
-                    } else {
-                        for i in chain_start..chain_end {
-                            let elem = {
-                                let arr_val = self.get_var(&chain.arr, 0)?;
-                                match arr_val {
-                                    Value::Array(elems) => elems.get(i as usize).cloned().unwrap_or(Value::Unknown(Type::Unknown)),
-                                    _ => return Err(RuntimeError::NotAContainer { pos: 0, name: chain.arr.clone() }),
-                                }
-                            };
-                            {
-                                let arr_val = self.get_var(&chain.arr, 0)?;
-                                if let Value::Array(mut elems) = arr_val {
-                                    if chain.start_slot < elems.len() { elems[chain.start_slot] = elem; }
-                                    self.set_var(chain.arr.clone(), Value::Array(elems), 0)?;
-                                }
-                            }
-                            self.push_scope();
-                            let mut should_break = false;
-                            for stmt in &chain.body {
-                                match self.run_semantic_stmt(stmt) {
-                                    Ok(_) => {}
-                                    Err(RuntimeError::BreakSignal) => { should_break = true; break; }
-                                    Err(RuntimeError::ContinueSignal) => break,
-                                    Err(e) => { self.pop_scope(); return Err(e); }
-                                }
-                            }
-                            self.pop_scope();
-                            if should_break { break; }
-                        }
-                    }
-                }
-
-                if let Some(result_expr) = result {
-                    self.eval_semantic_expr(result_expr)?;
-                }
-
-                Ok(())
-            }
-        }
-    }
-
-    fn semantic_value_to_runtime(&mut self, sv: &SemanticValue) -> Value {
-        match sv {
-            SemanticValue::Num(n) => Value::Num(*n),
-            SemanticValue::Float(f) => Value::Float(*f),
-            SemanticValue::Str(s) => {
-                let (off, len) = self.alloc_str(s);
-                Value::Str(off, len)
-            }
-            SemanticValue::Bool(b) => Value::Bool(*b),
-            SemanticValue::Char(c) => Value::Char(*c),
-            SemanticValue::EnumVariant { enum_name, variant_name, .. } => {
-                Value::EnumVariant(enum_name.clone(), variant_name.clone())
-            }
-            SemanticValue::Unknown => Value::Unknown(Type::T32),
-        }
-    }
-
-    fn apply_unary(&self, op: &Op, val: Value, pos: usize) -> Result<Value, RuntimeError> {
-        match (op, val) {
-            (Op::Minus, Value::Num(n)) => Ok(Value::Num(n.wrapping_neg())),
-            (Op::Minus, Value::Float(f)) => Ok(Value::Float(-f)),
-            (Op::Not, Value::Bool(b)) => Ok(Value::Bool(!b)),
-            (Op::Not, Value::TBool(n)) => Ok(Value::TBool(if n == 0 { 1 } else if n == 1 { 0 } else { 2 })),
-            (Op::Mul, Value::Array(elems)) => {
-                elems.get(0).cloned().ok_or_else(|| RuntimeError::BadAssignTarget { pos })
-            }
-            (Op::Mul, v) => Ok(v),
-            _ => Err(RuntimeError::TypeMismatch { pos, expected: Type::Unknown, got: Type::Unknown }),
-        }
-    }
-
-    pub fn call_semantic_func(&mut self, callee: &str, args: &[SemanticCallArg], pos: usize) -> Result<Value, RuntimeError> {
-        // Built-in dispatch — names come from the single-source-of-truth
-        // registry (crate::frontend::builtins, #008). Each kind's body is its
-        // pre-#008 behavior verbatim. Only `is_known` may fall out of the match
-        // (when its first arg is not an Expr) to the user-function path below;
-        // every other arm returns.
-        if let Some(def) = builtins::lookup(callee) {
-            match def.kind {
-                BuiltinKind::IsKnown => {
-                    if let Some(SemanticCallArg::Expr(e)) = args.first() {
-                        let val = self.eval_semantic_expr(e)?;
-                        return Ok(match val {
-                            Value::Unknown(_) | Value::TBool(2) => Value::Bool(false),
-                            _ => Value::Bool(true),
-                        });
-                    }
-                }
-
-                // exit(code) / exit() — request process termination. Returns the
-                // Exit control-flow signal (NOT a process::exit call here); the
-                // top-level loop / --test loop translates it to a real exit so
-                // stdout can be flushed first. Negative codes pass through to i32
-                // unchanged; the OS layer (POSIX) may truncate to u8 — not
-                // normalised.
-                BuiltinKind::Exit => {
-                    let code: i32 = match args.first() {
-                        None => 0,
-                        Some(SemanticCallArg::Expr(e)) => {
-                            let v = self.eval_semantic_expr(e)?;
-                            match v {
-                                Value::Num(n) => {
-                                    if n < i32::MIN as i128 || n > i32::MAX as i128 {
-                                        // Out of i32 range: surface as a normal
-                                        // runtime error, not an Exit signal.
-                                        return Err(RuntimeError::TypeMismatch {
-                                            pos,
-                                            expected: Type::T32,
-                                            got: Type::T64,
-                                        });
-                                    }
-                                    n as i32
-                                }
-                                other => {
-                                    return Err(RuntimeError::TypeMismatch {
-                                        pos,
-                                        expected: Type::T32,
-                                        got: type_of_value(&other),
-                                    });
-                                }
-                            }
-                        }
-                        // Non-Expr arg form (copy/copyfree/copyinto) is not valid
-                        // for exit; arity/shape is enforced in the semantic phase.
-                        Some(_) => 0,
-                    };
-                    return Err(RuntimeError::Exit(code));
-                }
-
-                // print (with newline) and println — see #034: both currently
-                // behave identically (no distinct newline).
-                BuiltinKind::Print | BuiltinKind::Println => {
-                    for arg in args {
-                        if let SemanticCallArg::Expr(e) = arg {
-                            let v = self.eval_semantic_expr(e)?;
-                            self.print_value(&v);
-                        }
-                    }
-                    return Ok(Value::Num(0));
-                }
-
-                // printn (no newline)
-                BuiltinKind::Printn => {
-                    for arg in args {
-                        if let SemanticCallArg::Expr(e) = arg {
-                            let v = self.eval_semantic_expr(e)?;
-                            self.print_value_inline(&v);
-                        }
-                    }
-                    return Ok(Value::Num(0));
-                }
-
-                // assert(cond) — runtime error if condition is false
-                BuiltinKind::Assert => {
-                    if let Some(SemanticCallArg::Expr(e)) = args.first() {
-                        let val = self.eval_semantic_expr(e)?;
-                        let passed = match val {
-                            Value::Bool(b) => b,
-                            Value::Num(n) => n != 0,
-                            _ => false,
-                        };
-                        if !passed {
-                            return Err(RuntimeError::AssertionFailed {
-                                msg: "assertion failed".to_string(),
-                                pos,
-                            });
-                        }
-                    }
-                    return Ok(Value::Num(0));
-                }
-
-                // assert_eq(a, b) — runtime error if a != b
-                BuiltinKind::AssertEq => {
-                    let mut iter = args.iter();
-                    let left = if let Some(SemanticCallArg::Expr(e)) = iter.next() {
-                        self.eval_semantic_expr(e)?
-                    } else { return Ok(Value::Num(0)); };
-                    let right = if let Some(SemanticCallArg::Expr(e)) = iter.next() {
-                        self.eval_semantic_expr(e)?
-                    } else { return Ok(Value::Num(0)); };
-                    let equal = match (&left, &right) {
-                        (Value::Num(a), Value::Num(b)) => a == b,
-                        (Value::Bool(a), Value::Bool(b)) => a == b,
-                        (Value::Float(a), Value::Float(b)) => a == b,
-                        (Value::Str(ao, al), Value::Str(bo, bl)) => {
-                            self.resolve_str(*ao, *al) == self.resolve_str(*bo, *bl)
-                        }
-                        (Value::ResultOk(a), Value::ResultOk(b)) => a == b,
-                        (Value::ResultErr(a), Value::ResultErr(b)) => a == b,
-                        _ => false,
-                    };
-                    if !equal {
-                        return Err(RuntimeError::AssertionFailed {
-                            msg: format!("assert_eq failed: {} != {}",
-                                value_to_string(self, left),
-                                value_to_string(self, right)),
-                            pos,
-                        });
-                    }
-                    return Ok(Value::Num(0));
-                }
-
-                // read(var) — reads a line from stdin into var
-                BuiltinKind::Read => {
-                    if let Some(SemanticCallArg::Expr(e)) = args.first() {
-                        let mut input = String::new();
-                        std::io::stdin().read_line(&mut input).unwrap_or(0);
-                        let input = input.trim_end_matches('\n').trim_end_matches('\r').to_string();
-                        let (off, len) = self.alloc_str(&input);
-                        if let SemanticExprKind::VarRef { name, .. } = &e.kind {
-                            self.set_var(name.clone(), Value::Str(off, len), 0)?;
-                        }
-                        return Ok(Value::Str(off, len));
-                    }
-                    return Ok(Value::Num(0));
-                }
-
-                // input("prompt", var) — prints prompt then reads into var
-                BuiltinKind::Input => {
-                    let mut iter = args.iter();
-                    // First arg is the prompt string
-                    if let Some(SemanticCallArg::Expr(prompt_expr)) = iter.next() {
-                        let prompt_val = self.eval_semantic_expr(prompt_expr)?;
-                        match &prompt_val {
-                            Value::Str(off, len) => print!("{}", self.resolve_str(*off, *len)),
-                            other => print!("{}", value_to_string(self, other.clone())),
-                        }
-                        use std::io::Write;
-                        std::io::stdout().flush().unwrap_or(());
-                    }
-                    // Second arg is the variable to fill
-                    if let Some(SemanticCallArg::Expr(var_expr)) = iter.next() {
-                        let mut input = String::new();
-                        std::io::stdin().read_line(&mut input).unwrap_or(0);
-                        let input = input.trim_end_matches('\n').trim_end_matches('\r').to_string();
-                        let (off, len) = self.alloc_str(&input);
-                        if let SemanticExprKind::VarRef { name, .. } = &var_expr.kind {
-                            self.set_var(name.clone(), Value::Str(off, len), 0)?;
-                        }
-                        return Ok(Value::Str(off, len));
-                    }
-                    return Ok(Value::Num(0));
-                }
-            }
-        }
-
-        // Native copy semantics — no fallback to AST path needed
-        let sem_func = self.semantic_funcs.get(callee).cloned()
-            .ok_or_else(|| RuntimeError::UndefinedVar { pos, name: callee.to_string() })?;
-
-        let outer_scope_idx = self.scopes.len() - 1;
-        let mut resolved: Vec<(BindingId, String, Value, Option<String>)> = Vec::new();
-        // (param binding, inner param name, value, bleed_back outer name if .copy)
-
-        for (param, arg) in sem_func.params.iter().zip(args.iter()) {
-            match arg {
-                SemanticCallArg::Expr(e) => {
-                    let val = self.eval_semantic_expr(e)?;
-                    resolved.push((param.binding, param.name.clone(), val, None));
-                }
-                SemanticCallArg::Copy { name, .. } => {
-                    let val = self.get_var(name, pos)?;
-                    resolved.push((param.binding, param.name.clone(), val, Some(name.clone())));
-                    // bleed_back registered below after push_function_scope
-                }
-                SemanticCallArg::CopyFree { name, .. } => {
-                    let val = self.get_var(name, pos)?;
-                    resolved.push((param.binding, param.name.clone(), val, None));
-                    // no bleed_back — isolated copy
-                }
-                SemanticCallArg::CopyInto(bindings) => {
-                    let mut map = HashMap::new();
-                    for b in bindings {
-                        let val = self.get_var(&b.name, pos)?;
-                        map.insert(b.name.clone(), val);
-                    }
-                    resolved.push((param.binding, param.name.clone(), Value::Container(map), None));
-                }
-            }
-        }
-
-        self.push_function_scope();
-
-        let result = (|| -> Result<Value, RuntimeError> {
-            for (pbinding, pname, val, bleed_outer) in resolved {
-                let ty = type_of_value(&val);
-                self.set_var_typed(pbinding, pname.clone(), ty, val, pos)?;
-                if let Some(outer_name) = bleed_outer {
-                    if let Some(frame) = self.scopes.last_mut() {
-                        frame.bleed_back.insert(pname, (outer_scope_idx, outer_name));
-                    }
-                }
-            }
-            for stmt in &sem_func.body {
-                match self.run_semantic_stmt(stmt) {
-                    Ok(_) => {}
-                    Err(RuntimeError::EarlyReturn(v)) => return Ok(v),
-                    Err(e) => return Err(e),
-                }
-            }
-            if let Some(expr) = &sem_func.ret_expr {
-                self.eval_semantic_expr(expr)
-            } else {
-                Ok(Value::Num(0))
-            }
-        })();
-
-        self.pop_scope();
-        result
-    }
-
-    fn call_semantic_method(&mut self, instance: &str, method: &str, args: &[SemanticCallArg], pos: usize) -> Result<Value, RuntimeError> {
-        // Get primary instance value and type
-        let instance_val = self.get_var(instance, pos)?;
-        let type_name = match &instance_val {
-            Value::Struct(name, _) => name.clone(),
-            _ => return Err(RuntimeError::NotAContainer { pos, name: instance.to_string() }),
-        };
-
-        // Look up in semantic impl registry
-        let (aliases, sem_func) = self.semantic_impls
-            .get(&(type_name.clone(), method.to_string()))
-            .cloned()
-            .ok_or_else(|| RuntimeError::UndefinedVar { pos, name: format!("{}.{}", instance, method) })?;
-
-        // Aliases beyond the first are passed as leading args at the call site
-        // args[0..extra_alias_count] = additional alias values
-        // args[extra_alias_count..] = regular params
-        let extra_alias_count = aliases.len().saturating_sub(1);
-
-        // Resolve additional alias values from leading args
-        let mut alias_vals: Vec<(BindingId, String, Type, Value)> = Vec::new();
-        // First alias is always the dot receiver
-        let (first_alias_binding, first_alias_name, first_alias_type) = &aliases[0];
-        alias_vals.push((*first_alias_binding, first_alias_name.clone(), first_alias_type.clone().into(), instance_val.clone()));
-
-        // Remaining aliases come from leading call args
-        for i in 0..extra_alias_count {
-            let (alias_binding, alias_name, alias_type) = &aliases[i + 1];
-            let val = match args.get(i) {
-                Some(SemanticCallArg::Expr(e)) => self.eval_semantic_expr(e)?,
-                Some(SemanticCallArg::Copy { name, .. }) => self.get_var(name, pos)?,
-                Some(SemanticCallArg::CopyFree { name, .. }) => self.get_var(name, pos)?,
-                _ => return Err(RuntimeError::UndefinedVar { pos, name: format!("missing alias arg {}", i) }),
-            };
-            alias_vals.push((*alias_binding, alias_name.clone(), alias_type.clone().into(), val));
-        }
-
-        // Resolve regular params from remaining args
-        let regular_args = &args[extra_alias_count..];
-        let mut resolved_params: Vec<(BindingId, String, Value)> = Vec::new();
-        for (param, arg) in sem_func.params.iter().zip(regular_args.iter()) {
-            let val = match arg {
-                SemanticCallArg::Expr(e) => self.eval_semantic_expr(e)?,
-                SemanticCallArg::Copy { name, .. } => self.get_var(name, pos)?,
-                SemanticCallArg::CopyFree { name, .. } => self.get_var(name, pos)?,
-                SemanticCallArg::CopyInto(bindings) => {
-                    let mut map = HashMap::new();
-                    for b in bindings {
-                        map.insert(b.name.clone(), self.get_var(&b.name, pos)?);
-                    }
-                    Value::Container(map)
-                }
-            };
-            resolved_params.push((param.binding, param.name.clone(), val));
-        }
-
-        // Push scope
-        self.push_function_scope();
-
-        let result = (|| -> Result<Value, RuntimeError> {
-            // Bind all aliases into scope
-            for (alias_binding, alias_name, alias_ty, val) in &alias_vals {
-                self.set_var_typed(*alias_binding, alias_name.clone(), alias_ty.clone(), val.clone(), pos)?;
-            }
-
-            // Bind regular params
-            for (pbinding, pname, val) in resolved_params {
-                let ty = type_of_value(&val);
-                self.set_var_typed(pbinding, pname, ty, val, pos)?;
-            }
-
-            // Run semantic body
-            for stmt in &sem_func.body {
-                match self.run_semantic_stmt(stmt) {
-                    Ok(_) => {}
-                    Err(RuntimeError::EarlyReturn(v)) => return Ok(v),
-                    Err(e) => return Err(e),
-                }
-            }
-
-            // Evaluate return expression
-            if let Some(expr) = &sem_func.ret_expr {
-                self.eval_semantic_expr(expr)
-            } else {
-                Ok(Value::Num(0))
-            }
-        })();
-
-        // Capture all alias mutations before popping scope
-        let mut mutated_aliases: Vec<(String, Value)> = Vec::new();
-        if result.is_ok() {
-            for (_, alias_name, _, _) in &alias_vals {
-                if let Ok(mutated) = self.get_var(alias_name, pos) {
-                    mutated_aliases.push((alias_name.clone(), mutated));
-                }
-            }
-        }
-
-        self.pop_scope();
-
-        // Write all alias mutations back to caller scope
-        // First alias writes back to instance
-        // Additional aliases write back to the variable names passed as leading args
-        if result.is_ok() {
-            for (i, (_, mutated)) in mutated_aliases.iter().enumerate() {
-                if i == 0 {
-                    let _ = self.set_var(instance.to_string(), mutated.clone(), pos);
-                } else if i <= extra_alias_count {
-                    // Get the original variable name from the leading arg
-                    if let Some(SemanticCallArg::Expr(e)) = args.get(i - 1) {
-                        if let SemanticExprKind::VarRef { name, .. } = &e.kind {
-                            let _ = self.set_var(name.clone(), mutated.clone(), pos);
-                        }
-                    } else if let Some(SemanticCallArg::Copy { name, .. }) = args.get(i - 1) {
-                        let _ = self.set_var(name.clone(), mutated.clone(), pos);
-                    }
-                }
-            }
-        }
-
-        result
-    }
-
-    fn expand_interpolation(&self, s: &str) -> String {
-        let mut result = String::new();
-        let mut chars = s.chars().peekable();
-        while let Some(c) = chars.next() {
-            if c == '{' {
-                let mut var_name = String::new();
-                let mut closed = false;
-                for inner in chars.by_ref() {
-                    if inner == '}' {
-                        closed = true;
-                        break;
-                    }
-                    var_name.push(inner);
-                }
-                if closed && !var_name.is_empty() {
-                    let val = self.scopes.iter().rev()
-                        .find_map(|frame| frame.get_by_name(&var_name))
-                        .and_then(|entry| entry.val.clone());
-                    match val {
-                        Some(v) => result.push_str(&value_to_string(self, v)),
-                        None => {
-                            result.push('{');
-                            result.push_str(&var_name);
-                            result.push('}');
-                        }
-                    }
-                } else if !closed {
-                    result.push('{');
-                    result.push_str(&var_name);
-                }
-            } else {
-                result.push(c);
-            }
-        }
-        result
-    }
-
-    fn print_value(&self, val: &Value) {
-        match val {
-            Value::Str(off, len) => {
-                let s = self.resolve_str(*off, *len);
-                println!("{}", self.expand_interpolation(s));
-            }
-            _ => println!("{}", value_to_string(self, val.clone())),
-        }
-    }
-
-    fn print_value_inline(&self, val: &Value) {
-        match val {
-            Value::Str(off, len) => {
-                let s = self.resolve_str(*off, *len);
-                print!("{}", self.expand_interpolation(s));
-            }
-            _ => print!("{}", value_to_string(self, val.clone())),
-        }
-    }
-
-    fn run_semantic_when(&mut self, val: Value, arms: &[SemanticWhenArm]) -> Result<Value, RuntimeError> {
-        for arm in arms {
-            let matches = match &arm.pattern {
-                SemanticWhenPattern::Literal(sv) => {
-                    let pat_val = self.semantic_value_to_runtime(sv);
-                    match (&val, &pat_val) {
-                        (Value::Str(vo, vl), Value::Str(po, pl)) => {
-                            self.resolve_str(*vo, *vl) == self.resolve_str(*po, *pl)
-                        }
-                        _ => val == pat_val
-                    }
-                }
-                SemanticWhenPattern::Range(lo, hi, inclusive) => {
-                    let lo_val = self.semantic_value_to_runtime(lo);
-                    let hi_val = self.semantic_value_to_runtime(hi);
-                    match (&val, &lo_val, &hi_val) {
-                        (Value::Num(v), Value::Num(l), Value::Num(h)) => {
-                            if *inclusive { v >= l && v <= h } else { v >= l && v < h }
-                        }
-                        _ => false,
-                    }
-                }
-                SemanticWhenPattern::EnumVariant { enum_name, variant_name, .. } => {
-                    match &val {
-                        Value::EnumVariant(e, v) => e == enum_name && v == variant_name,
-                        _ => false,
-                    }
-                }
-                SemanticWhenPattern::Catchall => true,
-            };
-            if matches {
-                self.push_scope();
-                let mut last_val = Value::Num(0);
-                for s in &arm.body {
-                    match s {
-                        SemanticStmt::ExprStmt { expr, .. } => {
-                            last_val = self.eval_semantic_expr(expr)?;
-                        }
-                        _ => {
-                            match self.run_semantic_stmt(s) {
-                                Ok(_) => {}
-                                Err(e) => { self.pop_scope(); return Err(e); }
-                            }
-                        }
-                    }
-                }
-                self.pop_scope();
-                return Ok(last_val);
-            }
-        }
-        Ok(Value::Num(0))
-    }
 }
 
-fn value_to_string(rt: &RunTime, v: Value) -> String {
+
+pub(crate) fn value_to_string(rt: &RunTime, v: Value) -> String {
     match v {
         Value::Num(n) => n.to_string(),
         Value::Float(x) => x.to_string(),
@@ -1885,7 +169,7 @@ fn value_to_string(rt: &RunTime, v: Value) -> String {
     }
 }
 
-fn apply_numeric_cast(val: Value, to: &SemanticType) -> Value {
+pub(crate) fn apply_numeric_cast(val: Value, to: &SemanticType) -> Value {
     match val {
         Value::Num(n) => {
             let truncated = match to {
@@ -1914,7 +198,7 @@ fn apply_numeric_cast(val: Value, to: &SemanticType) -> Value {
     }
 }
 
-fn type_of_value(v: &Value) -> Type {
+pub(crate) fn type_of_value(v: &Value) -> Type {
     match v {
         Value::Num(_) => Type::T128,
         Value::Float(_) => Type::F64,
@@ -1933,7 +217,7 @@ fn type_of_value(v: &Value) -> Type {
     }
 }
 
-fn value_matches_type(v: &Value, t: &Type) -> bool {
+pub(crate) fn value_matches_type(v: &Value, t: &Type) -> bool {
     match (v, t) {
         (Value::Num(_), Type::T8) => true,
         (Value::Num(_), Type::T16) => true,
@@ -1963,7 +247,7 @@ fn value_matches_type(v: &Value, t: &Type) -> bool {
     }
 }
 
-fn is_ident(s: &str) -> bool {
+pub(crate) fn is_ident(s: &str) -> bool {
     let mut chars = s.chars();
     let Some(first) = chars.next() else {
         return false;
@@ -1974,7 +258,7 @@ fn is_ident(s: &str) -> bool {
     chars.all(|c| c == '_' || c.is_alphanumeric())
 }
 
-fn as_f64(v: &Value) -> Option<f64> {
+pub(crate) fn as_f64(v: &Value) -> Option<f64> {
     match v {
         Value::Num(n) => Some(*n as f64),
         Value::Float(x) => Some(*x),
@@ -1982,7 +266,7 @@ fn as_f64(v: &Value) -> Option<f64> {
     }
 }
 
-fn expand_template(rt: &RunTime, s: &str, pos: usize) -> Result<String, RuntimeError> {
+pub(crate) fn expand_template(rt: &RunTime, s: &str, pos: usize) -> Result<String, RuntimeError> {
     let mut out = String::new();
     let mut chars = s.chars().peekable();
